chore: refactor sha validation
